Clean code [ADD] Destructor.c avec clean cfg [MV] fonction clean dans sh.h, en attente du fix debug pour debug les derniers problemes
