+++ srcs/evaluator/cleaner.c
@@ -1,24 +1,11 @@
 #include "exec.h"
 #include "libft.h"
-
-
 #include "sh.h"
-#include "ft_printf.h"
 
 void	free_fd(void *del, size_t u)
 {
 	(void)u;
 	free(del);
-}
-
-void	del_struct_tvar(void *del, size_t u)
-{
-	t_var	*v;
-
-	(void)u;
-	v = del;
-	ft_del_tab((void **)v->ctab);
-	free(v);
 }
 
 void	del_struct_process(void *del, size_t u)
@@ -53,18 +40,4 @@ void	del_struct_job(void *del, size_t u)
 {
 	routine_clean_job(del, u);
 	free(del);
-}
-
-void	clean_cfg(t_cfg *shell)
-{
-	ft_lstdel(&shell->env, del_struct_tvar);
-	ft_lstdel(&shell->intern, del_struct_tvar);
-	ft_lstdel(&shell->sp, del_struct_tvar);
-	ft_lstdel(&shell->job, del_struct_job);
-	ft_dlstdel(&shell->history);
-	ft_hash_del_map(&shell->map, free);
-	ft_hash_del_map(&shell->input_map, NULL);
-	if (shell->debug > 2)
-		close(shell->debug);
-	ft_bzero(shell, sizeof(t_cfg));
 }
