+++ srcs/evaluator/launcher_process.c
@@ -3,6 +3,7 @@
 #include "libft.h"
 #include "var.h"
 #include "sh.h"
+#include "ft_printf.h"
 #include "job_control.h"
 #include <unistd.h>
 
@@ -11,7 +12,6 @@
 #include <stdio.h> //perror !!!!]]
 #include <stdlib.h>
 
-#include "ft_printf.h"
 #include "sh.h"
 
 
@@ -57,7 +57,7 @@ uint8_t		builtin_process(t_job *j, t_process *p)
 	return (p->ret);
 }
 
-int		parent_process(t_job *job, t_process *process, int fd_pipe, char **envp)
+uint8_t		parent_process(t_job *job, t_process *process, int fd_pipe, char **envp)
 {
 	if (fd_pipe)
 		if (close(fd_pipe) == -1)
@@ -74,10 +74,10 @@ int		parent_process(t_job *job, t_process *process, int fd_pipe, char **envp)
 		if (tcsetpgrp(STDIN_FILENO, job->pgid))
 			perror("[PARENT PROCESS] error tcsetpgrp"); //perror
 	ft_del_tab((void **)envp);
-	return (0);
+	return (SUCCESS);
 }
 
-int		child_process(t_job *job, t_process *p, int fd_pipe, char **envp)
+uint8_t		child_process(t_job *job, t_process *p, int fd_pipe, char **envp)
 {
 	if (fd_pipe)
 		if (close(fd_pipe) == -1)
@@ -95,12 +95,11 @@ int		child_process(t_job *job, t_process *p, int fd_pipe, char **envp)
 			perror("[CHILD PROCESS] error tcsetpgrp");
 	do_pipe(p);
 	do_redir(p->fd);
-	if (p->setup & R_ERROR)
+	if (p->setup & ERROR)
 	{
 		ft_dprintf(STDERR_FILENO,"%s", p->message);
-		exit(1);
+		exit(p->ret);
 	}
-	process_errors_handling(p);
 	if (p->setup & BUILTIN)
 		exit(builtin_process(job, p));  //que faire de envp??????
 	if ((execve(p->path, p->av, envp)) == -1)
@@ -108,12 +107,11 @@ int		child_process(t_job *job, t_process *p, int fd_pipe, char **envp)
 	exit(1);
 }
 
-int		fork_process(t_job *job, t_process *p)
+uint8_t		fork_process(t_job *job, t_process *p)
 {
 	char	**envp;
 
 	envp = create_tab_var(p->env, 0); //problematique, a voir ac l'assignement
-	p->status = RUNNING;
 	if ((p->pid = fork()) == -1)
 		perror("fork:"); // perror
 	if (!(p->pid))
@@ -123,19 +121,10 @@ int		fork_process(t_job *job, t_process *p)
 	return (0);
 }
 
-void	close_fd(t_list *lst)
-{
-	int16_t	*fd;
-
-	fd = lst->data;
-	if (fd[0] > 2)
-		close(fd[0]);
-	if (fd[1] > 2)
-		close(fd[1]);
-}
-
 void	run_process(t_cfg *shell, t_job *j, t_process *p)
 {
+
+	p->status = RUNNING | (p->status & ~WAITING);
 	process_type(p);
 	process_assign(shell, p, p->assign); // not cmd != false
 	debug_print_process(j, p, "run_process");
@@ -156,42 +145,4 @@ void	run_process(t_cfg *shell, t_job *j, t_process *p)
 		fork_process(j, p);
 	ft_lstiter(p->fd, close_fd);
 	return ;
-}
-
-uint8_t		routine_ending_job(t_cfg *shell, t_job *job)
-{
-	char	*ret;
-
-	if (!shell->interactive)
-		wait_process(job);
-	else if (job->fg)
-	{
-	/*	tcsetpgrp(STDIN_FILENO, job->pgid) // seulememt ici du coup ?? */
-		wait_process(job);
-		tcsetpgrp(STDIN_FILENO, shell->pid);
-		set_termios(TCSADRAIN, &shell->term_origin);
-	}
-	else
-		set_job_background(job);
-	ret = ft_itoa(job->ret);
-	setvar_update(find_var(shell->sp, "?"), ret);
-	ft_strdel(&ret);
-	return (job->ret);
-}
-
-
-uint8_t		run_job(t_cfg *shell, t_job *job, t_list *process)
-{
-	job->status |= RUNNING;
-	ft_lstiter(job->process, job_redir);
-	while (process)
-	{
-		routine_process(shell, process, &job->pipe);
-		run_process(shell, job, process->data);
-		process = process->next;
-		if (job->pipe.tmp)
-			if (close(job->pipe.tmp) == -1)
-				perror("[check and do pipe] close error:");  //perror
-	}
-	return (routine_ending_job(shell, job));
 }
