+++ srcs/evaluator/tools_redir.c
@@ -4,7 +4,16 @@
 #include <unistd.h>
 #include <sys/stat.h>
 
-void	do_my_dup2(int16_t fd1, int16_t fd2)
+void		add_fd_process(t_list **fd, int16_t source, int16_t target)
+{
+	int16_t		tab[2];
+
+	tab[0] = source;
+	tab[1] = target;
+	ft_lst_push_back(fd, tab, sizeof(tab));
+}
+
+void		do_my_dup2(int16_t fd1, int16_t fd2)
 {
 	if (fd1 == -1)
 		return ;
@@ -12,13 +21,39 @@ void	do_my_dup2(int16_t fd1, int16_t fd2)
 		ft_ex("Fatal Error DUP2\nexit\n");
 }
 
-uint8_t		bad_fd(int fd)
+void		do_redir(t_list *fd)
 {
-	struct stat buf;
+	int16_t *t;
 
-	if (fstat(fd, &buf) == FALSE)
-		return (FAILURE);
-	return (SUCCESS);
+	while (fd)
+	{
+		t = fd->data;
+		if (t[0] == -1)
+			close(t[1]);
+		else if (t[0] != t[1])
+			do_my_dup2(t[0], t[1]);
+		fd = fd->next;
+	}
+}
+
+void		do_pipe(t_process *p)
+{
+	if (p->std[0] != STDIN_FILENO &&
+			p->std[0] != -1)
+	{
+		if (dup2(p->std[0], STDIN_FILENO) == -1)
+			perror("[do_pipe]1 dup2 error:");
+		if (close(p->std[0]) == -1)
+			perror("[do_pipe]1 close error:");
+	}
+	if (p->std[1] != STDOUT_FILENO &&
+			p->std[0] != -1)
+	{
+		if (dup2(p->std[1], STDOUT_FILENO) == -1)
+			perror("[do_pipe]2 dup2 error:");
+		if (close(p->std[1]) == -1)
+			perror("[do_pipe]2 close error:");
+	}
 }
 
 uint32_t	path_gearing(t_redir *r, char **path, int right)
